finished updating polars to 0.44

+++ src/num_ext/woe_iv.rs
@@ -30,14 +30,17 @@ fn get_woe_frame(discrete_col: &Series, target: &Series) -> PolarsResult<LazyFra
         .lazy()
         .drop_nulls(None)
         .group_by([col("value")])
-        .agg([len().alias("cnt"), col("target").sum().alias("goods")])
+        .agg([
+            len().cast(DataType::Float64).alias("cnt"), 
+            col("target").sum().cast(DataType::Float64).alias("goods")
+            ])
         .select([
             col("value"),
-            ((col("goods") + lit(1)).cast(DataType::Float64)
-                / (col("goods").sum() + lit(2)).cast(DataType::Float64))
+            ((col("goods") + lit(1f64))
+                / (col("goods").sum() + lit(2f64)))
             .alias("good_pct"),
-            ((col("cnt") - col("goods") + lit(1)).cast(DataType::Float64)
-                / (col("cnt").sum() - col("goods").sum() + lit(2)).cast(DataType::Float64))
+            ((col("cnt") - col("goods") + lit(1f64))
+                / (col("cnt").sum() - col("goods").sum() + lit(2f64)))
             .alias("bad_pct"),
         ])
         .with_column(
